injury_predictor/injury_predictor.py
from sklearn.linear_model import LogisticRegression

def train_injury_predictor(df):
    # Assuming df contains columns like 'previous_injuries', 'biomechanics_data'
    X = df.drop('injury', axis=1)
    y = df['injury']

    # Train injury prediction model
    model = LogisticRegression()
    model.fit(X, y)
    
    return model

def predict_injury_risk(model, athlete_data):
    # Predict injury risk based on input data
    risk_score = model.predict_proba(athlete_data)[:, 1]
    if risk_score > 0.7:
        return True
    return False

# Usage
# model = train_injury_predictor(training_data)
# injury_risk = predict_injury_risk(model, athlete_data)
