video_analysis/video_analysis.py
import cv2
import numpy as np
from posenet import load_model, detect_pose

def analyze_performance(video_path, model_path):
    # Load pre-trained PoseNet model
    model = load_model(model_path)

    # Initialize video capture
    cap = cv2.VideoCapture(video_path)
    while cap.isOpened():
        ret, frame = cap.read()
        if not ret:
            break

        # Detect key points
        keypoints = detect_pose(frame, model)

        # Analyze key points for specific performance insights
        insights = analyze_keypoints(keypoints)
        print(insights)

        # Display the frame
        cv2.imshow('Performance Analysis', frame)
        if cv2.waitKey(1) & 0xFF == ord('q'):
            break

    cap.release()
    cv2.destroyAllWindows()

def analyze_keypoints(keypoints):
    # Placeholder for custom analysis algorithms
    return {"analysis": "Detailed feedback based on keypoints"}

# Use this module with a video file and model to perform analysis
# analyze_performance('athlete_video.mp4', 'posenet.pth')
