requirements.txt
opencv-python-headless
posenet-pytorch
numpy
pandas
scikit-learn
openai
